fix(LJ-570): Remove bad characters in Azure Storage view instead of reverse proxy
The reverse proxy doesn't seem to be used when directly accessing a
/cloud-media link. Moving this logic to the view ensures that file names
are always cleaned on GET.
